cosine_similarity
Cosine Similarity is a metric used to measure the similarity between two vectors. This metric determines similarity by calculating the angle between the vectors. The closer the value is to 1, the more similar the vectors (and thus the texts they represent) are. The closer the value is to 0, the more dissimilar the vectors are.